improve ci for esp32s2
