+++ .github/workflows/build.yml
@@ -53,12 +53,21 @@ jobs:
       uses: actions/setup-node@v1
 
     - name: Cache MSP430 Toolchain
+      uses: actions/cache@v2
       id: cache-msp430
-      uses: actions/cache@v1
       with:
         path: /tmp/dl/
         # Increment gcc version number at end when updating downloads
-        key: msp430-${{ runner.os }}-9.2.0.50
+        key: ${{ runner.os }}-msp430-9.2.0.50
+
+    - name: Install MSP430 Toolchain
+      if: steps.cache-msp430.outputs.cache-hit != 'true'
+      env:
+        MSP430GCC_URL: http://software-dl.ti.com/msp430/msp430_public_sw/mcu/msp430/MSPGCC/9_2_0_0/export/msp430-gcc-9.2.0.50_linux64.tar.bz2      
+      run: |
+        mkdir -p /tmp/dl/
+        [ -f "/tmp/dl/msp430-gcc.tar.bz2" ] || wget --progress=dot:mega $MSP430GCC_URL -O /tmp/dl/msp430-gcc.tar.bz2
+        tar -C $HOME -xaf /tmp/dl/msp430-gcc.tar.bz2
 
     - name: Install Toolchains
       env:
@@ -72,9 +81,6 @@ jobs:
         echo `echo $HOME/opt/xPacks/@xpack-dev-tools/riscv-none-embed-gcc/*/.content/bin` >> $GITHUB_PATH
 
         # TI MSP430 GCC
-        mkdir -p /tmp/dl/
-        [ -f "/tmp/dl/msp430-gcc.tar.bz2" ] || wget --progress=dot:mega $MSP430GCC_URL -O /tmp/dl/msp430-gcc.tar.bz2
-        tar -C $HOME -xaf /tmp/dl/msp430-gcc.tar.bz2
         echo `echo $HOME/msp430-gcc-*_linux64/bin` >> $GITHUB_PATH
 
     - name: Checkout TinyUSB
@@ -113,17 +119,7 @@ jobs:
     - name: Setup Node.js
       uses: actions/setup-node@v1
 
-    - name: Cache MSP430 Toolchain
-      id: cache-msp430
-      uses: actions/cache@v1
-      with:
-        path: /tmp/dl/
-        # Increment gcc version number at end when updating downloads
-        key: msp430-${{ runner.os }}-9.2.0.50
-
     - name: Install Toolchains
-      env:
-        MSP430GCC_URL: http://software-dl.ti.com/msp430/msp430_public_sw/mcu/msp430/MSPGCC/9_2_0_0/export/msp430-gcc-9.2.0.50_linux64.tar.bz2
       run: |
         # ARM & RISC-V GCC from xpack
         npm install --global xpm
@@ -131,12 +127,6 @@ jobs:
         xpm install --global @xpack-dev-tools/riscv-none-embed-gcc@latest
         echo `echo $HOME/opt/xPacks/@xpack-dev-tools/arm-none-eabi-gcc/*/.content/bin` >> $GITHUB_PATH
         echo `echo $HOME/opt/xPacks/@xpack-dev-tools/riscv-none-embed-gcc/*/.content/bin` >> $GITHUB_PATH
-
-        # TI MSP430 GCC
-        mkdir -p /tmp/dl/
-        [ -f "/tmp/dl/msp430-gcc.tar.bz2" ] || wget --progress=dot:mega $MSP430GCC_URL -O /tmp/dl/msp430-gcc.tar.bz2
-        tar -C $HOME -xaf /tmp/dl/msp430-gcc.tar.bz2
-        echo `echo $HOME/msp430-gcc-*_linux64/bin` >> $GITHUB_PATH
 
     - name: Checkout TinyUSB
       uses: actions/checkout@v2
@@ -153,13 +143,15 @@ jobs:
     - name: Build
       run: python3 tools/build_family.py ${{ matrix.family }}
 
-  # Build ESP32S
-  build-esp32s:
+  # Build ESP32S2
+  build-esp32s2:
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
-        example: ['board_test', 'cdc_msc_freertos', 'hid_composite_freertos']
+        board:
+        - 'espressif_kaluga_1'
+        - 'espressif_saola_1'
 
     steps:
     - name: Setup Python
@@ -174,4 +166,4 @@ jobs:
         submodules: 'false'
 
     - name: Build
-      run: docker run --rm -v $PWD:/project -w /project espressif/idf:latest python3 tools/build_esp32s.py ${{ matrix.example }}
+      run: docker run --rm -v $PWD:/project -w /project espressif/idf:latest python3 tools/build_esp32s2.py ${{ matrix.board }}
